tests/wpt/tests/css/css-overflow/parsing/webkit-line-clamp-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Overflow: parsing -webkit-line-clamp with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-overflow/#webkit-line-clamp">
<meta name="assert" content="-webkit-line-clamp supports only the grammar 'none | <integer>'.">
<meta name="assert" content="Zero or negative max-lines integers are invalid.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("-webkit-line-clamp", 'auto');

test_invalid_value("-webkit-line-clamp", '0');
test_invalid_value("-webkit-line-clamp", '-5');

test_invalid_value("-webkit-line-clamp", 'none "~"');
test_invalid_value("-webkit-line-clamp", '1 "~"');
</script>
</body>
</html>
